Warp generation (#36)
* Add DynamicWarps to World

* Reverse externals direction on generated worlds

* Add dynamic warps to worlds

* Populate warps in the generated world

* Add test for warp generation

* Add index to dynamic warp

* Add intention processing logging

* Add headers to logs

* Restore behaviourMap in PrimordiaVille

+++ Shared/LuceRPG.Models/WorldObject.fs
@@ -106,7 +106,7 @@ module WorldObject =
     /// Time taken by the object to move one square
     let travelTime (obj: Payload): int64 =
         match obj.t with
-        | Type.Player _ -> System.TimeSpan.FromMilliseconds(float(250)).Ticks
+        | Type.Player _ -> System.TimeSpan.FromMilliseconds(float(150)).Ticks
         | Type.NPC _ -> System.TimeSpan.FromMilliseconds(float(280)).Ticks
         | _ -> 0L
